Move useTracks into own file
