+++ src/utils/crud/track.js
@@ -1,30 +1,3 @@
-import {useState, useEffect} from 'react'
-
-export function useTracks(channelId, database) {
-	const [tracks, setTracks] = useState([])
-	const [error, setError] = useState(null)
-
-	useEffect(() => {
-		const fetchData = async () => {
-			let res
-			try {
-				res = await database
-					.from('channel_track')
-					.select('id:track_id, created_at, track_id(url, title, description)')
-					.eq('channel_id', channelId)
-				setTracks(res.data)
-				setError(null)
-			} catch (error) {
-				setError(error)
-				console.log('Error fetching tracks', error)
-			}
-		}
-		fetchData()
-	}, [channelId, database])
-
-	return {data: tracks, error}
-}
-
 export const createTrack = async ({database, track, channelId, userId}) => {
 	const {url, title, description} = track
 
@@ -32,7 +5,6 @@ export const createTrack = async ({database, track, channelId, userId}) => {
 	const res = await database.from('tracks').insert({url, title, description}).single()
 
 	// Avoid touching next query (return early) if it did not succeed.
-	// console.log('creating track', res)
 	if (res.error) return res
 
 	// Create junction row
